requirements.txt
django
django-filter
django-rest-framework
djoser
django-cors-headers
djangorestframework_simplejwt
gunicorn
psycopg2
django-excel
pymongo
pyexcel-xlsx